+++ policyengine_canada/parameters/gov/cra/benefits/gis_spa/widows_eligibility_age.yaml
@@ -1,0 +1,8 @@
+description: Age at which a widow becomes eligible for the GIS
+values:
+  2020-01-01: 60
+metadata:
+  label: SPA Widow Eligibility Age
+  reference:
+  - title: Government of Canada - Guaranteed Income Supplement
+    href: https://www.canada.ca/en/services/benefits/publicpensions/cpp/old-age-security/guaranteed-income-supplement/allowance-survivor.html
